// src/modules/visit-record/interfaces/visit-record.interface.ts
export interface IVisitRecord {
  date: Date;
}

export interface ITotalVisits{
  point_name: string
  point_id: string
  age_baby: number
  age_small_kid: number
  age_preescholar_kid: number
  age_primary_kid: number
  age_adolescent: number
  age_young_adult: number
  age_early_adulthood: number
  age_middle_adult: number
  age_mayor_adult: number
  entnicithy_afroecuadorian: number
  entnicithy_halfbreed: number
  entnicithy_native: number
  entnicithy_white: number
  isPregnant: number
  hasLittleChildren: number
  total_under_age: number
  total_adults: number
  total_man: number
  total_woman: number
  total: number
}